Moved ConfigValidationException to main package. Fixed some visibility modifiers.
